% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordenadas.R
\name{canton_crtm05}
\alias{canton_crtm05}
\title{Busca coordenadas de cantones en formato CRTM05}
\usage{
canton_crtm05(canton)
}
\arguments{
\item{canton}{}
}
\value{
coordenadas Coordenadas del canton en WSG84-GPS
}
\description{
El sistema de coordenadas geograficas usadas por el gobierno de Costa Rica
es CRTM05. Esta función devuelve estas coordenadas tal como se publicaron
en
}
\details{
La Gaceta No. 100 del 26 de mayo 2009
DIVISIÓN TERRITORIAL ADMINISTRATIVA DE COSTA RICA
SEGÚN DECRETO N° 35213-MG PUBLICADO EN
LA GACETA No 85 DEL 5 DE MAYO DE 2009
}
