% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordenadas.R
\name{canton_wsg84}
\alias{canton_wsg84}
\title{Busca coordenadas de cantones en formato WSG84-GPS}
\usage{
canton_wsg84(canton)
}
\arguments{
\item{canton}{}
}
\value{
coordenadas Coordenadas del canton en WSG84-GPS
}
\description{
Para poder crear mapas con, por ejemplo, Leaflet necesitamos coordenadas
sobre el sistema WSG84-GPS, pero los datos oficiales de costa rica usan
CRTM05. Esta función devuelve datos en WSG84-GPS.
}
