% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordenadas.R
\name{crgeo_canton}
\alias{crgeo_canton}
\title{Muestra datos disponibles para los cantones}
\usage{
crgeo_canton(canton = "todo")
}
\arguments{
\item{canton}{}
}
\description{
Muestra los datos que tenemos para cantones. Sin argumentos (o con el
argumento "todo") la función va mostrar todos los datos disponibles.
}
