% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordenadas.R
\name{crgeo_distrito}
\alias{crgeo_distrito}
\title{Muestra datos disponibles para los distritos}
\usage{
crgeo_distrito(distrito = "todo")
}
\arguments{
\item{canton}{}
}
\value{
coordenadas
}
\description{
Muestra los datos que tenemos para distritos. Sin argumentos (o con el
argumento "todo") la función va mostrar todos los datos disponibles.
}
