% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordenadas.R
\name{distrito_crtm05}
\alias{distrito_crtm05}
\title{Busca coordenadas de distritos en formato CRTM05}
\usage{
distrito_crtm05(canton, unidad_territorial)
}
\arguments{
\item{canton}{el nombre del cantón}

\item{unidad_territorial}{el nombre de la unidad territorial}
}
\value{
coordenadas Coordenadas del distrito en WSG84-GPS
}
\description{
El sistema de coordenadas geograficas usadas por el gobierno de Costa Rica
es CRTM05. Esta función devuelve estas coordenadas tal como se publicaron
en
}
\details{
La Gaceta No. 100 del 26 de mayo 2009
DIVISIÓN TERRITORIAL ADMINISTRATIVA DE COSTA RICA
SEGÚN DECRETO N° 35213-MG PUBLICADO EN
LA GACETA No 85 DEL 5 DE MAYO DE 2009
}
